requirements.txt
# 仅做 pyannote.audio 说话人分割（时间段 + 说话人标签）
pyannote.audio>=3.1.0
torch
torchaudio
faster-whisper
# FastAPI 接口
fastapi
uvicorn[standard]
python-multipart
requests
websockets
pydub
# LLM 流水线（推断说话人、纠错）
python-dotenv
langchain-openai
langchain-core
